[TASK] Move some `AtRulBlockList` tests to unit tests
Part of #758
