+++ tests/CSSList/AtRuleBlockListTest.php
@@ -5,10 +5,7 @@ declare(strict_types=1);
 namespace Sabberworm\CSS\Tests\CSSList;
 
 use PHPUnit\Framework\TestCase;
-use Sabberworm\CSS\Comment\Commentable;
-use Sabberworm\CSS\CSSList\AtRuleBlockList;
 use Sabberworm\CSS\Parser;
-use Sabberworm\CSS\Renderable;
 use Sabberworm\CSS\Settings;
 
 /**
@@ -49,36 +46,6 @@ final class AtRuleBlockListTest extends TestCase
                 ',
             ],
         ];
-    }
-
-    /**
-     * @test
-     */
-    public function implementsAtRule(): void
-    {
-        $subject = new AtRuleBlockList('');
-
-        self::assertInstanceOf(AtRuleBlockList::class, $subject);
-    }
-
-    /**
-     * @test
-     */
-    public function implementsRenderable(): void
-    {
-        $subject = new AtRuleBlockList('');
-
-        self::assertInstanceOf(Renderable::class, $subject);
-    }
-
-    /**
-     * @test
-     */
-    public function implementsCommentable(): void
-    {
-        $subject = new AtRuleBlockList('');
-
-        self::assertInstanceOf(Commentable::class, $subject);
     }
 
     /**
